+++ SchoolOrganizer/src/main/java/com/myapp/repository/RoleDao.java
@@ -6,6 +6,8 @@ import java.util.List;
 
 public interface RoleDao {
 
+    void setUpDefaultRole(Long id);
+
     List<Role> getRolesByUserEmail(String email);
 
     Role getRoleByName(String name);
